+++ sign/policy.go
@@ -4,9 +4,9 @@ package sign
 // and the number of participants.
 type ParticipationMask interface {
 	// CountEnabled returns the number of participants
-	CountEnabled() int
+	CountEnabled() uint32
 	// CountTotal returns the number of candidates
-	CountTotal() int
+	CountTotal() uint32
 }
 
 // Policy represents a fully customizable cosigning policy deciding what
@@ -35,16 +35,16 @@ func (p CompletePolicy) Check(m ParticipationMask) bool {
 // least the given threshold number of participants t have cosigned to make a
 // collective signature valid.
 type ThresholdPolicy struct {
-	thold int64
+	thold uint32
 }
 
 // NewThresholdPolicy returns a new ThresholdPolicy with the given threshold.
-func NewThresholdPolicy(thold int) *ThresholdPolicy {
-	return &ThresholdPolicy{thold: int64(thold)}
+func NewThresholdPolicy(thold uint32) *ThresholdPolicy {
+	return &ThresholdPolicy{thold: thold}
 }
 
 // Check verifies that at least a threshold number of participants have
 // contributed to a collective signature.
 func (p ThresholdPolicy) Check(m ParticipationMask) bool {
-	return int64(m.CountEnabled()) >= p.thold
+	return m.CountEnabled() >= p.thold
 }
